+++ data/hts/reference.py
@@ -10,8 +10,11 @@ for calibration and validation of the simulation output.
 """
 
 def execute(context):
-    df_trips = context.stage("data.hts.selected")[2]
-    df_trips.to_csv("%s/reference.csv" % context.config("output_path"), sep = ";", index = False)
+    data = context.stage("data.hts.selected")
+
+    data[0].to_csv("%s/hts_households.csv" % context.config("output_path"), sep = ";", index = False)
+    data[1].to_csv("%s/hts_persons.csv" % context.config("output_path"), sep = ";", index = False)
+    data[2].to_csv("%s/hts_trips.csv" % context.config("output_path"), sep = ";", index = False)
 
 def validate(context):
     output_path = context.config("output_path")